requirements.txt
tensorflow-gpu==1.10
keras==2.1.5
h5py
numpy
opencv-python
sklearn
matplotlib